+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Order templates directories by title in `dg-templates-listing` view.
   [sgeulette]
+- Improved `DXDocumentGenerationHelperView.get_value` by adding optional obj parameter.
+  [sgeulette]
 
 3.35 (2022-08-26)
 -----------------